fix: broken typings for exported command and plugin
BREAKING CHANGE: rename/simplify import path
