+++ package.json
@@ -2,8 +2,17 @@
   "name": "cypress-image-diff-js",
   "version": "0.0.0",
   "description": "Visual regression testing tool with cypress",
-  "main": "dist/command.js",
-  "types": "types/index.d.ts",
+  "exports": {
+    "./command": {
+      "types": "./types/command.d.ts",
+      "default": "./dist/command"
+    },
+    "./plugin": {
+      "types": "./types/plugin.d.ts",
+      "default": "./dist/plugin"
+    }
+  },
+  "files": ["dist", "types"],
   "bin": {
     "cypress-image-diff": "./bin/cypress-image-diff.js"
   },
