+++ types/command.d.ts
@@ -1,0 +1,12 @@
+
+import type { CompareSnapshotOptions } from './options'
+
+declare namespace Cypress {
+  interface Chainable<Subject = any> {
+    compareSnapshot(options: string | CompareSnapshotOptions): Chainable<Element>
+  }
+}
+
+declare function compareSnapshotCommand(): void
+
+export default compareSnapshotCommand;
