+++ types/plugin.d.ts
@@ -1,0 +1,8 @@
+/// <reference types="cypress" />
+
+declare function getCompareSnapshotsPlugin(
+  on: Cypress.PluginEvents,
+  config: Cypress.PluginConfigOptions,
+): void
+
+export default getCompareSnapshotsPlugin;
